replace low-level bitmap handling code with glide
The code that is in place currently decodes a bitmap on the main thread,
which can cause application responsiveness issues. The code also doesn't
appear to scale down the bitmap in cases where the bitmap is much larger
than the ImageView size. These two issues are described here:
https://developer.android.com/topic/performance/graphics/index.html

It looks like they recommend the use of Glide to handle these problems, so I've replaced
the lower-level code with Glide code. This code may not be correct. I've never
used glide (I typically use Picasso).